Move scoping functions to a separate ScopeGuards header.
