+++ src/libs/renderer/renderer/gl_helpers.h
@@ -1,15 +1,12 @@
 #pragma once
 
+
 #include "GL_Loader.h"
-#include "Query.h"
-
-#include <handy/Guard.h>
-
-#include <math/Rectangle.h>
 
 
 namespace ad {
 namespace graphics {
+
 
     // Sadly, the glGen* symbols imported by the GL loader
     // are not compile time constants
@@ -21,6 +18,7 @@ namespace graphics {
 //    return name;
 //}
 
+
 inline GLuint reserve(void(GLAPIENTRY * aGlGenFunction)(GLsizei, GLuint *))
 {
     GLuint name;
@@ -31,7 +29,6 @@ inline GLuint reserve(void(GLAPIENTRY * aGlGenFunction)(GLsizei, GLuint *))
 
 namespace detail 
 {
-
 
 template <class T_resource>
 class NameBase
@@ -58,7 +55,6 @@ private:
     GLuint mResource;
 };
 
-
 } // namespace detail
 
 
@@ -74,84 +70,6 @@ class Name : public detail::NameBase<T_resource>
 public:
     using detail::NameBase<T_resource>::NameBase;
 };
-
-
-// TODO should we use the GLenum directly instead?
-// It should be in BufferBase, but ScopedBind needs it atm
-enum class BufferType
-{
-    Array = GL_ARRAY_BUFFER,
-    ElementArray = GL_ELEMENT_ARRAY_BUFFER,
-    Uniform = GL_UNIFORM_BUFFER,
-};
-
-
-// Forward declaration
-template <BufferType>
-struct Buffer;
-
-
-// TODO might become a single template function once all other stuff is made generic
-class ScopedBind
-{
-public:
-    template <class T_resource, class... VT_args>
-    ScopedBind(const T_resource & aResource, VT_args &&... aArgs) :
-        mGuard{[previous = getBound(aResource, aArgs...), aArgs...]()
-               { bind(previous, aArgs...); }}
-    {
-        bind(aResource, aArgs...);
-    }
-
-    // TODO scoped bind for Indexed version
-
-private:
-    Guard mGuard;
-};
-
-
-inline Guard scopeFeature(GLenum aFeature, bool aEnable)
-{
-    bool wasEnabled = isEnabled(aFeature);
-
-    auto handler = [aFeature](bool enable)
-    {
-        if(enable) glEnable(aFeature);
-        else glDisable(aFeature);
-    };
-    handler(aEnable);
-
-    return Guard{ std::bind(handler, wasEnabled) };
-}
-
-
-inline Guard scopeDepthMask(bool aEnable)
-{
-    bool wasEnabled = isEnabled(GL_DEPTH_WRITEMASK);
-
-    auto handler = [](bool enable)
-    {
-        glDepthMask(enable);
-    };
-    handler(aEnable);
-
-    return Guard{ std::bind(handler, wasEnabled) };
-}
-
-
-inline Guard scopeViewport(math::Rectangle<GLint> aViewport)
-{
-    std::array<GLint, 4> previous;
-    glGetIntegerv(GL_VIEWPORT, previous.data());
-
-    glViewport(aViewport.xMin(), aViewport.yMin(),
-               aViewport.width(), aViewport.height());
-
-    return Guard{ [previous]()
-        {
-            glViewport(previous[0], previous[1], previous[2], previous[3]);
-        }};
-}
 
 
 } // namespace graphics
